# _workspace/03_design/api_openapi.yaml
openapi: 3.1.0
info:
  title: Hoban JHA Agent API
  version: 0.2.0
  description: |
    호반그룹 LLM/RAG 기반 작업위험성평가(JHA) 지원 에이전트 PoC API.
    계약 우선(Contract-first). frontend·ERP·eval 병렬 개발 기준.

    ## 핵심 규칙
    - 세션 상태 머신: CREATED→CLASSIFIED→ASSESSED→(PENDING_REVIEW→REVIEWED)?→FINALIZED→REGISTERING→COMPLETED/REGISTER_FAILED
    - 곱16 경계셀(강도4×빈도4): human_review_required=true, critical_register="O (잠정)".
      안전관리자 review 확정 전 finalize 호출 시 409 REVIEW_REQUIRED.
    - result_type: ok | low_confidence | no_match | refused_partial | refused_full
    - LLM 응답은 jha_generation_template JSON 스키마를 코드가 재검증(등급 재계산·인용검증 citations⊆retrieved).
    - 외부 LLM 호출 전 화이트리스트 필드만 전송(data_security_policy.md). PII 의심 필드 마스킹 강제.

    ## 버저닝
    - 경로 prefix `/v1`. breaking change → `/v2`. 비파괴 추가는 minor(info.version).
    - deprecation: Sunset 헤더 + 최소 1 릴리스 유지.
  contact:
    name: backend-engineer (Hoban JHA PoC)
servers:
  - url: https://jha-api.hoban.internal
    description: 사내 운영(예정)
  - url: http://localhost:8000
    description: 로컬 개발
...
security:
  - bearerAuth: []
paths:
...
  /v1/kb/rows:
    get:
      tags: [kb]
      operationId: listKbRows
      summary: KB 행 검색·필터·페이징 (safety_manager+)
      description: |
        전사 하위공종 위험요인 운영 저장소(SSOT) 조회. q(위험요인·대책·세부항목 텍스트),
        분류·등급·중점등록 필터, offset/limit 페이징, sort.
      parameters:
        - { name: q, in: query, required: false, schema: { type: string } }
        - { name: major_type, in: query, required: false, schema: { type: string } }
        - { name: sub_type, in: query, required: false, schema: { type: string } }
        - { name: accident_type, in: query, required: false, schema: { type: string } }
        - { name: risk_grade, in: query, required: false, schema: { type: string, enum: ["상","중","하"] } }
        - { name: critical_register, in: query, required: false, schema: { type: string, enum: ["O","X"] } }
        - { name: include_deleted, in: query, required: false, schema: { type: boolean, default: false } }
        - { name: offset, in: query, required: false, schema: { type: integer, default: 0, minimum: 0 } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 50, minimum: 1, maximum: 500 } }
        - { name: sort, in: query, required: false, schema: { type: string, default: chunk_id } }
      responses:
        '200':
          description: 행 목록
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbRowList' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [kb]
      operationId: createKbRow
      summary: KB 행 생성 (safety_manager+)
      description: |
        신규 행 생성. chunk_id 는 N{seq} 자동 부여(기존 R##### 보존).
        **서버 강제 도메인 규칙**: risk_grade 는 강도×빈도 임계곱 재계산(입력 무시),
        중점등록은 등급'상'⇔O 자동(곱16 경계셀만 입력 존중 + boundary_cell 플래그).
        분류 검증: 대/중공종은 taxonomy_lookup 존재값 필수, 신규 세부항목은 허용 + is_new_detail.
        생성 후 자동 재인덱싱(디바운스) 스케줄 → 핫스왑.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KbRowWrite' }
      responses:
        '201':
          description: 생성됨(서버 재계산 등급·중점등록 포함)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbRow' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /v1/kb/rows/{chunk_id}:
    parameters:
      - { name: chunk_id, in: path, required: true, schema: { type: string } }
    get:
      tags: [kb]
      operationId: getKbRow
      summary: KB 행 단건 (safety_manager+)
      responses:
        '200':
          description: 단건
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbRow' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [kb]
      operationId: updateKbRow
      summary: KB 행 수정 (safety_manager+)
      description: 동일 도메인 규칙·검증 적용. 수정 후 자동 재인덱싱.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KbRowWrite' }
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbRow' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [kb]
      operationId: deleteKbRow
      summary: KB 행 soft delete (safety_manager+)
      description: row_status=deleted. 인덱스에서 제외(재인덱싱 반영). 인용 시 "삭제된 행" 표기.
      responses:
        '200':
          description: 삭제됨(soft)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbRow' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/kb/stats:
    get:
      tags: [kb]
      operationId: kbStats
      summary: KB 통계·재인덱싱 상태 (safety_manager+)
      description: 행수·분류 카운트 + reindex_status·index_version·last_reindex_at·doc_count·regression_recommended.
      responses:
        '200':
          description: 통계
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KbStats' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/kb/reindex:
    post:
      tags: [kb]
      operationId: kbReindex
      summary: 수동 전체 재인덱싱 트리거 (safety_manager+)
      description: 디바운스 없이 즉시 동기 재인덱싱 + 핫스왑. 실패 시 구 인덱스 유지(무중단).
      responses:
        '200':
          description: 재인덱싱 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReindexAck' }
        '401': { $ref: '#/components/responses/AuthError' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 호반 SSO(OIDC) 또는 PoC JWT. role claim(worker|safety_manager|admin) 포함.
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                에러 코드 체계(prefix):
                AUTH_* (401/403), VALIDATION_* (400/422),
                SESSION_* (404/409), RAG_* (200+result_type 또는 5xx),
                LLM_* (502/503), ERP_* (202/5xx), SECURITY_* (500 비노출),
                INTERNAL_* (500), RATE_* (429)
            message: { type: string, description: 사용자 표시용 메시지(한국어) }
            details:
              type: object
              additionalProperties: true
              description: 디버그 정보(사용자 비표시 가능)
            request_id: { type: string }

    # ── KB(지식베이스) CRUD 스키마 ──────────────────────────────────────────
    KbRowWrite:
      type: object
      required: [major_type, sub_type, detail_item, severity, frequency, hazard_text]
      description: |
        KB 행 생성/수정 입력. **risk_grade·critical_register 는 서버가 강제 재계산**하므로
        입력해도 무시된다(곱16 경계셀의 critical_register 만 존중). 클라이언트는 강도·빈도만 보낸다.
      properties:
        major_type: { type: string, description: 대공종(taxonomy_lookup 존재값) }
        sub_type: { type: string, description: 중공종(taxonomy_lookup 존재값) }
        detail_item: { type: string, description: 세부항목(신규 허용 → is_new_detail=true) }
        accident_type: { type: string, default: 기타, description: 재해형태(21종 enum) }
        severity: { type: integer, minimum: 1, maximum: 5 }
        frequency: { type: integer, minimum: 1, maximum: 5 }
        hazard_text: { type: string, minLength: 1 }
        controls: { type: string, default: "" }
        critical_register:
          type: string
          enum: ["O", "X", "O (잠정)"]
          nullable: true
          description: 곱16 경계셀에서만 존중. 그 외 서버 자동(상⇔O).
        legal_refs: { type: array, items: { type: string } }
    KbRow:
      type: object
      required: [chunk_id]
      properties:
        chunk_id: { type: string, description: "기존 R##### 또는 신규 N{seq}" }
        source_row: { type: integer, nullable: true, description: "신규 행은 음수(lineage 구분)" }
        major_type: { type: string }
        sub_type: { type: string }
        detail_item: { type: string }
        accident_type: { type: string }
        severity: { type: integer }
        frequency: { type: integer }
        risk_product: { type: integer, description: "강도×빈도(서버 계산)" }
        risk_grade: { type: string, enum: ["상","중","하"], description: "서버 재계산(임계곱)" }
        critical_register: { type: string, enum: ["O","X"], description: "서버 자동(상⇔O)" }
        boundary_cell: { type: boolean, description: "곱16(강도4×빈도4) 경계셀" }
        is_new_detail: { type: boolean, description: "taxonomy 미등록 신규 세부항목" }
        hazard_text: { type: string }
        hazard_items: { type: array, items: { type: string } }
        controls: { type: string }
        controls_items: { type: array, items: { type: string } }
        legal_refs: { type: array, items: { type: string } }
        row_status: { type: string, enum: ["active","deleted"] }
        updated_at: { type: string, nullable: true }
        updated_by: { type: string, nullable: true }
    KbRowList:
      type: object
      required: [rows, total, offset, limit]
      properties:
        rows: { type: array, items: { $ref: '#/components/schemas/KbRow' } }
        total: { type: integer }
        offset: { type: integer }
        limit: { type: integer }
    KbStats:
      type: object
      properties:
        active_rows: { type: integer }
        deleted_rows: { type: integer }
        new_rows: { type: integer, description: "N{seq} 신규 행 수" }
        by_major_type: { type: object, additionalProperties: { type: integer } }
        by_risk_grade: { type: object, additionalProperties: { type: integer } }
        reindex_status: { type: string, enum: ["idle","pending","running"] }
        index_version: { type: integer, description: "재인덱싱 증가 카운터" }
        last_reindex_at: { type: string, nullable: true }
        doc_count: { type: integer, description: "현재 인덱스 문서 수" }
        last_change_ratio: { type: number }
        regression_recommended: { type: boolean, description: "변경비율>5% → 회귀 권고(sync.py 정합)" }
    ReindexAck:
      type: object
      properties:
        status: { type: string }
        index_version: { type: integer }
        doc_count: { type: integer }
        last_reindex_at: { type: string, nullable: true }
        last_duration_ms: { type: number, nullable: true }
        regression_recommended: { type: boolean }
  responses:
    ValidationError:
      description: 입력 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: VALIDATION_FIELD_INVALID, message: "work_description은 2자 이상이어야 합니다." }
    AuthError:
      description: 인증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: AUTH_TOKEN_INVALID, message: "인증 토큰이 유효하지 않습니다." }
    Forbidden:
      description: 권한 부족
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: AUTH_INSUFFICIENT_ROLE, message: "이 작업을 수행할 권한이 없습니다." }
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: SESSION_NOT_FOUND, message: "세션을 찾을 수 없습니다." }
...